+++ src/pgc.erl
@@ -22,6 +22,7 @@
          query/2, query/3, query/4]).
 
 -export_type([pool_id/0,
+              error/0, notice/0,
               query_options/0, query_result/0, exec_result/0,
               column_name/0, row/0,
               oid/0, float_value/0,
@@ -34,6 +35,9 @@
 
 -type pool_id() :: atom().
 
+-type error() :: pgc_proto:error_and_notice_fields().
+-type notice() :: pgc_proto:error_and_notice_fields().
+
 -type query_options() :: #{column_names_as_atoms => boolean(),
                            rows_as_hashes => boolean()}.
 
@@ -41,12 +45,10 @@
                          [column_name()],
                          [row()],
                          NbAffectedRows :: non_neg_integer()}
-                      | {error,
-                         pgc_proto:error_and_notice_fields()}.
+                      | {error, error()}.
 -type exec_result() :: {ok,
                         NbAffectedRows :: non_neg_integer()}
-                     | {error,
-                        pgc_proto:error_and_notice_fields()}.
+                     | {error, error()}.
 
 -type column_name() :: binary() | atom().