+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 % erl-pgc changelog
 
 # Next Version
+## Features
+- Introduce the `pgc:error/0` and `pgc:notice/0` type to avoid manipulating
+  the `pgc_proto:error_and_notice_fields()` type.
 
 # 1.3.0
 ## Features